Initial attempt at cleanup
Divided all the handlers into separate classes. Unsure if it works
properly still, I have yet to test it. However, I need to commit all the
changes that I've made so far.
